refactor: move protocol.Addr to tnet.Addr
